[refactor | Misc] added Test ROms as well as moved some depreciated files

+++ src/gui/qt_tetroidnes/Game_Display.cpp
@@ -82,7 +82,7 @@ void GameDisplay::on_crt_shader_changed(const bool b)
         draw_func = [this](sf::Drawable &drawable)
         {
             crt_shader->setUniform("time", (float)(time_between_draw_ms * 0.001f));
-            crt_shader->setUniform("density", (float)1.9);
+            crt_shader->setUniform("density", 1.9f);
             crt_shader->setUniform("opacityScanline", 0.2f);
             crt_shader->setUniform("opacityNoise", 0.2f);
             crt_shader->setUniform("curvature", 7.5f);
